requirements.txt
argcomplete==1.10.3
beautifulsoup4==4.8.2
chardet==3.0.4
compressed-rtf==1.0.6
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
IMAPClient==2.1.0
lxml==4.9.2
olefile==0.46
packaging==23.1
pdf2image==1.16.3
pdfminer.six==20191110
Pillow==9.5.0
pycryptodome==3.18.0
pytesseract==0.3.10
python-pptx==0.6.21
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.8.1
textract==1.6.5
tzlocal==5.0.1
xlrd==1.2.0
XlsxWriter==3.1.2
